+++ crates/nostr/src/nips/mod.rs
@@ -44,7 +44,6 @@ pub mod nip49;
 pub mod nip51;
 pub mod nip53;
 pub mod nip56;
-#[cfg(feature = "nip57")]
 pub mod nip57;
 pub mod nip58;
 #[cfg(feature = "nip59")]